# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Febrafar - Planner",
        "description": "Febrafar planner",
        "version": "1.0"
    },
    "paths": {
...
        "/api/tasks/matches": {
            "get": {
                "tags": [
                    "Tasks"
                ],
                "summary": "Find all tasks that match certain criteria",
                "operationId": "0d118205fba2323f134181802c24e4cd",
                "parameters": [
                    {
                        "name": "title",
                        "in": "query",
                        "description": "Title to match",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "status",
                        "in": "query",
                        "description": "Status to match",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "pending",
                                "completed",
                                "in_progress"
                            ]
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of matching tasks successfully retrieved",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Task"
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                }
            }
        },
        "/api/tasks": {
            "get": {
                "tags": [
                    "Tasks"
                ],
                "summary": "Retrieve the list of tasks",
                "operationId": "df5acf0421229afb5de9a188daa35471",
                "responses": {
                    "200": {
                        "description": "List of all tasks successfully retrieved",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Task"
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                }
            },
            "post": {
                "tags": [
                    "Tasks"
                ],
                "summary": "Create a new task",
                "operationId": "9334fbf1876a35605faa8fe8f5aac896",
                "requestBody": {
                    "description": "Task creation data",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "title": {
                                        "description": "Title of the new task",
                                        "type": "string"
                                    },
                                    "description": {
                                        "description": "Description of the new task",
                                        "type": "string"
                                    },
                                    "deadline": {
                                        "description": "Deadline of the new task",
                                        "type": "string",
                                        "format": "date"
                                    },
                                    "start_date": {
                                        "description": "Start date of the new task",
                                        "type": "string",
                                        "format": "date"
                                    },
                                    "status": {
                                        "description": "Status of the new task",
                                        "type": "string",
                                        "enum": [
                                            "pending",
                                            "completed",
                                            "in_progress"
                                        ]
                                    },
                                    "task_type_id": {
                                        "description": "ID of the task type for the new task",
                                        "type": "string",
                                        "format": "uuid"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Task successfully created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Task"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "422": {
                        "description": "Invalid input data"
                    }
                }
            }
        },
        "/api/tasks/{task}": {
            "get": {
                "tags": [
                    "Tasks"
                ],
                "summary": "Retrieve a single task",
                "operationId": "88df9eee8686b4d2f72e0b1243c99f4d",
                "parameters": [
                    {
                        "name": "task",
                        "in": "path",
                        "description": "ID of the task to retrieve",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Task successfully retrieved",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Task"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "404": {
                        "description": "Task not found"
                    }
                }
            },
            "put": {
                "tags": [
                    "Tasks"
                ],
                "summary": "Update an existing task",
                "operationId": "f89899b0cb719dbbbfbe2ed6cf0239de",
                "parameters": [
                    {
                        "name": "task",
                        "in": "path",
                        "description": "ID of the task to update",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Updated task data",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "title": {
                                        "description": "Updated title of the task",
                                        "type": "string"
                                    },
                                    "description": {
                                        "description": "Updated description of the task",
                                        "type": "string"
                                    },
                                    "deadline": {
                                        "description": "Updated deadline of the task",
                                        "type": "string",
                                        "format": "date"
                                    },
                                    "start_date": {
                                        "description": "Updated start date of the task",
                                        "type": "string",
                                        "format": "date"
                                    },
                                    "status": {
                                        "description": "Updated status of the task",
                                        "type": "string",
                                        "enum": [
                                            "pending",
                                            "completed",
                                            "in_progress"
                                        ]
                                    },
                                    "task_type_id": {
                                        "description": "Updated ID of the task type for the task",
                                        "type": "string",
                                        "format": "uuid"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Task successfully updated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Task"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "404": {
                        "description": "Task not found"
                    },
                    "422": {
                        "description": "Invalid input data"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Tasks"
                ],
                "summary": "Delete an existing task",
                "operationId": "06c2d1c987e5949d7b1aaa42e4be391d",
                "parameters": [
                    {
                        "name": "task",
                        "in": "path",
                        "description": "ID of the task to delete",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Task successfully deleted"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "404": {
                        "description": "Task not found"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "Task": {
                "properties": {
                    "id": {
                        "description": "Task ID",
                        "type": "string"
                    },
                    "title": {
                        "description": "Task title",
                        "type": "string"
                    },
                    "description": {
                        "description": "Task description",
                        "type": "string"
                    },
                    "deadline": {
                        "description": "Task deadline",
                        "type": "string",
                        "format": "date"
                    },
                    "start_date": {
                        "description": "Task start date",
                        "type": "string",
                        "format": "date"
                    },
                    "status": {
                        "description": "Task status",
                        "type": "string",
                        "enum": [
                            "pending",
                            "completed",
                            "in_progress"
                        ]
                    },
                    "task_type_id": {
                        "description": "Task Type ID",
                        "type": "string",
                        "format": "uuid"
                    }
                },
                "type": "object"
            },
...
        }
    },
...
}